System automation

+++ src/main/java/com/example/MedicalWebInput/Controllers/PatientController.java
@@ -1,12 +1,11 @@
 package com.example.MedicalWebInput.Controllers;
 
-import com.example.MedicalWebInput.Data.PatientDto.CreatePatientDto;
-import com.example.MedicalWebInput.Data.PatientDto.PatientDto;
-import com.example.MedicalWebInput.Data.PatientDto.PatientLoginDto;
+import com.example.MedicalWebInput.Data.PatientDto.*;
 import com.example.MedicalWebInput.Models.Patient;
+import com.example.MedicalWebInput.Models.Photo;
 import com.example.MedicalWebInput.Repository.DrugRepository;
-import com.example.MedicalWebInput.Services.PatientService;
-import com.example.MedicalWebInput.Services.ScheduleService;
+import com.example.MedicalWebInput.Services.*;
+import com.google.firebase.messaging.FirebaseMessagingException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -18,6 +17,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 import javax.validation.constraints.NotNull;
+import java.io.IOException;
 
 @Controller
 @RequestMapping("/patient")
@@ -28,7 +28,11 @@ public class PatientController {
     @Autowired
     private ScheduleService scheduleService;
     @Autowired
-    private DrugRepository drugRepository;
+    private ReminderService reminderService;
+    @Autowired
+    private EmailService emailService;
+    @Autowired
+    private PhotoService photoService;
 
 //    *************************************************************************
 //    GetMappings
@@ -47,6 +51,20 @@ public class PatientController {
         return "Patient/patient_listing";
     }
 
+    @GetMapping("/profile_photo")
+    public String addProfilePhoto(@NotNull Model model, @NotNull HttpServletRequest httpServletRequest) {
+        HttpSession session = httpServletRequest.getSession();
+        if (session == null || session.getAttribute("patient_info") == null) {
+            // If the user is not logged in, redirect them to the login page
+            return "redirect:/patient/login";
+        }
+        BasicPatientDto patient = (BasicPatientDto) session.getAttribute("patient_info");
+        AddPhotoDto addPhotoDto = new AddPhotoDto();
+        addPhotoDto.setPatientId(patient.getId());
+        model.addAttribute("photo_data", addPhotoDto);
+        return "Patient/profile_photo";
+    }
+
     @GetMapping("/new_patient")
     public String addNewPatient(@NotNull Model model) {
         model.addAttribute("patient_data", new CreatePatientDto());
@@ -54,7 +72,12 @@ public class PatientController {
     }
 
     @GetMapping("/edit_drug/{drugId}")
-    public String editDrugInfo(@PathVariable Long drugId, @NotNull Model model) {
+    public String editDrugInfo(@PathVariable Long drugId, @NotNull Model model, HttpServletRequest httpServletRequest) {
+        HttpSession session = httpServletRequest.getSession();
+        if (session == null || session.getAttribute("patient_info") == null) {
+            // If the user is not logged in, redirect them to the login page
+            return "redirect:/patient/login";
+        }
         if (patientService.getDrugAndScheduleInfo(drugId) != null) {
             model.addAttribute("schedule_info", patientService.getDrugAndScheduleInfo(drugId));
             return "Drug/drug_edit";
@@ -64,7 +87,12 @@ public class PatientController {
     }
 
     @GetMapping("/delete_drug/{drugId}")
-    public String deleteDrugById(@PathVariable Long drugId) {
+    public String deleteDrugById(@PathVariable Long drugId, HttpServletRequest httpServletRequest) {
+        HttpSession session = httpServletRequest.getSession();
+        if (session == null || session.getAttribute("patient_info") == null) {
+            // If the user is not logged in, redirect them to the login page
+            return "redirect:/patient/login";
+        }
         Long patientId = patientService.getDrugInfo(drugId).getPatientId();
         patientService.deleteDrugById(drugId);
         return "redirect:/patient/patient_info";
@@ -75,6 +103,7 @@ public class PatientController {
         model.addAttribute("login_detail", new PatientLoginDto());
         return "Patient/patient_login";
     }
+
     @GetMapping("/logout")
     public String logoutPatient(@NotNull HttpServletResponse response, HttpSession session) {
         session.invalidate();
@@ -92,26 +121,38 @@ public class PatientController {
 
     @GetMapping("/forgot_password")
     public String forgotPassword(@NotNull Model model) {
+
         model.addAttribute("patient_email", new PatientLoginDto());
         return "Patient/patient_reset";
     }
 
     @GetMapping("/patient_info")
-    public String getPatientInfo(@NotNull Model model, HttpServletRequest httpServletRequest) {
+    public String getPatientInfo(@NotNull Model model, HttpServletRequest httpServletRequest) throws FirebaseMessagingException, IOException {
         HttpSession session = httpServletRequest.getSession();
-        if (session == null || session.getAttribute("patient_info") == null ) {
+        if (session == null || session.getAttribute("patient_info") == null) {
             // If the user is not logged in, redirect them to the login page
             return "redirect:/patient/login";
         }
-        PatientDto patient = (PatientDto) session.getAttribute("patient_info");
+        BasicPatientDto patient = (BasicPatientDto) session.getAttribute("patient_info");
         model.addAttribute("patient_data", patientService.getPatientById(patient.getId()));
         model.addAttribute("drug_info", patientService.getDrugByPatientId(patient.getId()));
         model.addAttribute("drug_present", patientService.checkDrug(patient.getId()));
-        model.addAttribute("schedule_info",scheduleService.getScheduleByPatientId(patient.getId()));
+        model.addAttribute("schedule_info", scheduleService.getScheduleByPatientId(patient.getId()));
         model.addAttribute("schedule_present", scheduleService.checkIfNull(patient.getId()));
-        model.addAttribute("stock_info",scheduleService.getStockInfo(patient.getId()));
-        model.addAttribute("stock_present",scheduleService.checkStock(patient.getId()));
-        model.addAttribute("test_data",patientService.getAllPatientTests(patient.getId()));
+        model.addAttribute("stock_info", scheduleService.getStockInfo(patient.getId()));
+        model.addAttribute("stock_present", scheduleService.checkStock(patient.getId()));
+        model.addAttribute("test_data", patientService.getAllPatientTests(patient.getId()));
+        Photo patientPhoto = photoService.getPhotoInfoByPatientId(patient.getId());
+
+        if (patientPhoto != null) {
+            String image = photoService.getImage(patientPhoto.getId());
+            model.addAttribute("image", image);
+        }
+
+        reminderService.setPatientId(patient.getId());
+        if (!patientService.getDrugByPatientId(patient.getId()).isEmpty()) {
+            reminderService.sendDoseReminders();
+        }
 
         return "HomePage";
     }
@@ -167,8 +208,23 @@ public class PatientController {
             redirectAttributes.addFlashAttribute("reset_error", "Email doesn't exist");
             return "redirect:/patient/forgot_password";
         }
-        String newPassword = patientService.setNewPassword(patientLoginDto.getEmail());
-        return newPassword;
+        String password = patientService.setNewPassword();
+        emailService.sendSimpleMessage(patientLoginDto.getEmail(), "Password Reset", "New Password: " + password);
+        patientService.changePassword(patientLoginDto.getEmail(), password);
+        redirectAttributes.addFlashAttribute("change_success", true);
+        return "redirect:/patient/login";
+    }
+
+    @PostMapping("/upload")
+    public String handleFileUpload(@ModelAttribute AddPhotoDto addPhotoDto, @NotNull RedirectAttributes redirectAttributes) throws IOException {
+        photoService.checkForCurrentPhoto(addPhotoDto.getPatientId());
+        Photo photo = photoService.addNewPhoto(addPhotoDto);
+        String image = photoService.getImage(photo.getId());
+        if (image.equals(null)) {
+            redirectAttributes.addFlashAttribute("upload_error", "Check your uploaded photo");
+            return "redirect:/patient/profile_photo/" + photo.getId();
+        }
+        return "redirect:/patient/patient_info";
     }
